Created basis for tooltips
-Created a tooltips system that utilizes unity events. Custom tooltip scripts can be written to display info dependant on what the mouse is hovered over.

-Setting up the events happens in UIHoverOverManager, while functionality for what is displayed is defined in the HoverTip.

This is currently just for object placement buttons, and individual data sets are provided by passing each placeable object type scriptable object to the SetToolTipData function in ToolTip.cs.

For nation data, the ToolTip class should be inherited in a new NationToolTip class  and override the SetToolTipData function and define what data is extrapolated from each nation for the tool tip. This script should then be placed on each nation object