Added nation tooltip
Added functionality for nation tool tip, as NationToolTip is placed on each object it can get desired data from each country and display it on the tooltip.

For displaying nation object based data e.g. number of tree projects. Each time an object is placed it is added to the nationObjectsPlaced list, in the tile.cs. As every nation has a tile script as a component, as well as a NationToolTip script, we can extrapolate data from that list in the nation tooltip.