Basis of timeManager implemented
A new singleton class has been added called TimeManager, this aims to abstract the functionality of the time system and allow this script to update the necessary systems or objects that depend on month passed data.

The new script TimeManager.cs handles updating the list of active growing trees each month, the growth time of each tree is handled in the TreeGrowth.cs and minuses 1 every time a month passes, called from the TimeManager. However, this value is not persistent, when save is hit the time remaining from the growth script is got and used to update the respective trees growth time remaining persistent data.