Purchase system implemented
The purchase system has been implemented for buying objects to place in the world, e.g. trees or infrastructure.

I have also refactored some of my systems, using enums for objectType instead of comparing object tag within if statements.

The purchase system is within the BuildingManager.cs, it has a reference to the player script with player data in it, when trying to purchase and place an item it checks to see if the player has more money than the object cost, if so then place and - the value of object from player money value